adding newly added files
